score modifier refactor

+++ src/main/java/com/sbh/bpm/controller/MasterProjectController.java
@@ -23,6 +23,7 @@ import javax.ws.rs.core.UriInfo;
 
 import com.google.gson.Gson;
 import com.sbh.bpm.model.ActivityName;
+import com.sbh.bpm.model.ExerciseScoreModifier;
 import com.sbh.bpm.model.MasterCertificationType;
 import com.sbh.bpm.model.MasterCriteria;
 import com.sbh.bpm.model.MasterCriteriaBlocker;
@@ -30,10 +31,12 @@ import com.sbh.bpm.model.MasterDocument;
 import com.sbh.bpm.model.MasterEvaluation;
 import com.sbh.bpm.model.MasterExercise;
 import com.sbh.bpm.model.MasterLevel;
+import com.sbh.bpm.model.MasterScoreModifier;
 import com.sbh.bpm.model.MasterTemplate;
 import com.sbh.bpm.model.MasterVendor;
 import com.sbh.bpm.model.ProjectAssessment;
 import com.sbh.bpm.service.IActivityNameService;
+import com.sbh.bpm.service.IExerciseScoreModifierService;
 import com.sbh.bpm.service.IMasterCertificationTypeService;
 import com.sbh.bpm.service.IMasterCriteriaBlockerService;
 import com.sbh.bpm.service.IMasterCriteriaService;
@@ -41,6 +44,7 @@ import com.sbh.bpm.service.IMasterDocumentService;
 import com.sbh.bpm.service.IMasterEvaluationService;
 import com.sbh.bpm.service.IMasterExerciseService;
 import com.sbh.bpm.service.IMasterLevelService;
+import com.sbh.bpm.service.IMasterScoreModifierService;
 import com.sbh.bpm.service.IMasterTemplateService;
 import com.sbh.bpm.service.IMasterVendorService;
 import com.sbh.bpm.service.IProjectAssessmentService;
@@ -78,6 +82,12 @@ public class MasterProjectController extends GcsUtil{
 
   @Autowired
   private IMasterExerciseService masterExerciseService;
+
+  @Autowired
+  private IMasterScoreModifierService masterScoreModifierService;
+
+  @Autowired
+  private IExerciseScoreModifierService exerciseScoreModifierService;
 
   @Autowired
   private IMasterLevelService masterLevelService;
@@ -593,6 +603,15 @@ public class MasterProjectController extends GcsUtil{
 
       return Response.status(400).entity(json).build();
     }
+
+    List<MasterScoreModifier> score_modifiers = (List<MasterScoreModifier>) masterScoreModifierService.findByMasterExerciseID(exerciseId);
+    if(score_modifiers.size()>0) {
+      Map<String, String> map = new HashMap<String, String>();
+      map.put("message", "Cannot delete, there is any master score modifier refer this subject, delete them first");
+      String json = new Gson().toJson(map);
+
+      return Response.status(400).entity(json).build();
+    }
     
     boolean status = masterExerciseService.deleteById(exerciseId);
     return Response.status(status ? 200 : 400).build();
@@ -625,7 +644,6 @@ public class MasterProjectController extends GcsUtil{
     masterExercise.setExerciseType(exercise.getExerciseType());
     masterExercise.setCode(exercise.getCode());
     masterExercise.setName(exercise.getName());
-    masterExercise.setScoreModifier(exercise.getScoreModifier());
     if (masterExercise.getExerciseType().equals("prequisite")) {
       masterExercise.setMaxScore(null);
     } else {
@@ -678,7 +696,10 @@ public class MasterProjectController extends GcsUtil{
   ) {
     List<MasterCriteriaBlocker> blockers = (List<MasterCriteriaBlocker>) masterCriteriaBlockerService.findBymasterCriteriaID(criteriaId);
     if(blockers.size()>0) {
-      return Response.status(400, "Cannot delete, there is any master blocker refer this subject, delete them first").build();
+      Map<String, String> map = new HashMap<String, String>();
+      map.put("message", "Cannot delete, there is any master blocker refer this subject, delete them first");
+      String json = new Gson().toJson(map);
+      return Response.status(400).entity(json.toString()).build();
     }
 
     List<MasterDocument> documents = (List<MasterDocument>) masterDocumentService.findBymasterCriteriaID(criteriaId);
@@ -734,6 +755,95 @@ public class MasterProjectController extends GcsUtil{
     masterCriteria = masterCriteriaService.save(masterCriteria);
 
     String json = new Gson().toJson(masterCriteria);
+    return Response.ok(json).build();
+  }
+
+  @GET
+  @Path(value = "/exercises/{exercise_id}/score_modifiers")
+  @Produces(MediaType.APPLICATION_JSON)
+  public Response allScoreModifiersByExercise(@HeaderParam("Authorization") String authorization, @PathParam("exercise_id") Integer exerciseID) { 
+    List<MasterScoreModifier> criterias = (List<MasterScoreModifier>) masterScoreModifierService.findByMasterExerciseID(exerciseID);
+
+    String json = new Gson().toJson(criterias);
+    return Response.ok(json).build();
+  }
+
+  @GET
+  @Path(value = "/score_modifiers")
+  @Produces(MediaType.APPLICATION_JSON)
+  public Response allMasterScoreModifier(@HeaderParam("Authorization") String authorization) {      
+    List<MasterScoreModifier> score_modifiers = (List<MasterScoreModifier>) masterScoreModifierService.findAll();
+
+    String json = new Gson().toJson(score_modifiers);
+    return Response.ok(json).build();
+  }
+
+  @GET
+  @Path(value = "/score_modifiers/{score_modifier_id}")
+  @Produces(MediaType.APPLICATION_JSON)
+  public Response GetMasterScoreModifier(@HeaderParam("Authorization") String authorization, @PathParam("score_modifier_id") Integer score_modifier_id) {      
+    MasterScoreModifier score_modifier = masterScoreModifierService.findById(score_modifier_id);
+
+    String json = new Gson().toJson(score_modifier);
+    return Response.ok(json).build();
+  }
+
+  @DELETE
+  @Path(value = "/score_modifiers/{score_modifier_id}")
+  @Produces(MediaType.APPLICATION_JSON)
+  public Response ScoreModifierDeletion(
+    @HeaderParam("Authorization") String authorization,
+    @PathParam("score_modifier_id") Integer scoreModifierId
+  ) {
+    List<ExerciseScoreModifier> blockers = (List<ExerciseScoreModifier>) exerciseScoreModifierService.findByMasterScoreModifierID(scoreModifierId);
+    if(blockers.size()>0) {
+      Map<String, String> map = new HashMap<String, String>();
+      map.put("message", "Cannot delete, there is any master blocker refer this subject, delete them first");
+      String json = new Gson().toJson(map);
+      return Response.status(400).entity(json).build();
+    }
+
+    boolean status = masterScoreModifierService.deleteById(scoreModifierId);
+    return Response.status(status ? 200 : 400).build();
+  }
+
+  @POST
+  @Path(value = "/score_modifiers")
+  @Produces(MediaType.APPLICATION_JSON)
+  @Consumes(MediaType.APPLICATION_JSON)
+  public Response saveScoreModifiers(@HeaderParam("Authorization") String authorization,
+                                MasterScoreModifier scoreModifier) {
+    List<MasterScoreModifier> masters = (List<MasterScoreModifier>) masterScoreModifierService.findByMasterExerciseID(scoreModifier.getMasterExerciseID());
+      if(masters.size()>0) {
+      Map<String, String> map = new HashMap<String, String>();
+      map.put("message", "Cannot Add more than 1 score modifier");
+      String json = new Gson().toJson(map);
+      return Response.status(400).entity(json).build();
+    }
+
+    scoreModifier.setCreatedAt(new Date());             
+    scoreModifier = masterScoreModifierService.save(scoreModifier);
+
+    String json = new Gson().toJson(scoreModifier);
+    return Response.ok(json).build();
+  }
+
+  @PATCH
+  @Path(value = "/score_modifiers/{score_modifier_id}")
+  @Produces(MediaType.APPLICATION_JSON)
+  @Consumes(MediaType.APPLICATION_JSON)
+  public Response editScoreModifiers(@HeaderParam("Authorization") String authorization,
+                                MasterScoreModifier score_modifier, @PathParam("score_modifier_id") Integer scoreModifierId) {
+    MasterScoreModifier masterScoreModifier = masterScoreModifierService.findById(scoreModifierId);
+    masterScoreModifier.setMasterExerciseID(score_modifier.getMasterExerciseID());
+    masterScoreModifier.setTitle(score_modifier.getTitle());
+    masterScoreModifier.setDescription(score_modifier.getDescription());
+    masterScoreModifier.setScoreModifier(score_modifier.getScoreModifier());
+    masterScoreModifier.setActive(score_modifier.getActive());
+
+    masterScoreModifier = masterScoreModifierService.save(masterScoreModifier);
+
+    String json = new Gson().toJson(masterScoreModifier);
     return Response.ok(json).build();
   }
 
@@ -999,7 +1109,10 @@ public class MasterProjectController extends GcsUtil{
                                ActivityName activityName) {   
     ActivityName act = (ActivityName) activityNameService.findById(activityNameId);
     if (act == null) {
-      return Response.status(400, "project document activityName not found").build();
+      Map<String, String> map = new HashMap<String, String>();
+      map.put("message", "project document activity name not found");
+      String json = new Gson().toJson(map);
+      return Response.status(400).entity(json.toString()).build();
     }
 
     act.setName(activityName.getName());
